Move whitelist files into video and notes
