+++ src/main/java/com/revature/chronicle/daos/VideoRepo.java
@@ -14,13 +14,24 @@ import java.util.Optional;
  */
 @Repository
 public interface VideoRepo extends JpaRepository<Video, Integer> {
-
+	//WHITELIST QUERY STRINGS
+	public static final String ADD_USER = "INSERT INTO video_whitelist VALUES ? ?";
+	public static final String GET_USERS = "SELECT * FROM video_whitelist WHERE video_id = ?";
+	
       /**
        * Finds videos with an offset and limit used for paging
        * @param offset where to begin
        * @param limit how many records should be returned
        * @return the videos
        */
-      @Query(value = "select * from video v order by v.date asc offset ?1 fetch next ?2 rows only",nativeQuery = true)
-      List<Video> findVideosWithOffsetAndLimit(int offset,int limit);
+	//VIDEO QUERIES
+    @Query(value = "select * from video v order by v.date asc offset ?1 fetch next ?2 rows only",nativeQuery = true)
+    List<Video> findVideosWithOffsetAndLimit(int offset,int limit);
+      
+    //WHITELIST QUERIES
+    @Query(value = ADD_USER, nativeQuery = true)
+    void addUser(int noteId, String userId);
+  	
+  	@Query(value = GET_USERS, nativeQuery = true)
+    List<String> getUsers(int noteId);
 }
