+++ src/main/java/com/revature/chronicle/services/VideoService.java
@@ -1,7 +1,9 @@
 package com.revature.chronicle.services;
 
 import com.revature.chronicle.daos.VideoRepo;
+import com.revature.chronicle.daos.VideoWhitelistRepo;
 import com.revature.chronicle.models.Tag;
+import com.revature.chronicle.models.User;
 import com.revature.chronicle.models.Video;
 import com.revature.chronicle.security.FirebaseInitializer;
 import org.slf4j.Logger;
@@ -108,6 +110,17 @@ public class VideoService {
             logger.warn(e.getMessage());
             return false;
         }
+    }
+    
+    //WHITELIST METHODS
+	public void addUserToWhitelist(Video video, List<User> users) {
+    	for(User user : users) {
+    		this.addUserToWhitelist(video, user);
+    	}
+    }
+    public void addUserToWhitelist(Video video, User user) {
+    	logger.info("Adding user to video whitelist" );
+    	videoRepo.addUser(video.getId(), user.getUserID());
     }
 
 //    //this method is to test the service method (use the repo method instead!)
